backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-common==1.1.28
azure-core==1.34.0
azure-search-documents==11.5.2
beautifulsoup4==4.13.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.9.0
neo4j==5.28.1
openai==1.78.1
playwright==1.52.0
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
pytz==2025.2
requests==2.29.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==1.26.20
uvicorn==0.34.2
